// src/generated/proto/stream_list.ts
export interface LiveChatMessage {
  kind?: string | undefined;
  etag?: string | undefined;
  id?: string | undefined;
  snippet?: LiveChatMessageSnippet | undefined;
  authorDetails?: LiveChatMessageAuthorDetails | undefined;
}

export interface LiveChatMessageAuthorDetails {
  channelId?: string | undefined;
  channelUrl?: string | undefined;
  displayName?: string | undefined;
  profileImageUrl?: string | undefined;
  isVerified?: boolean | undefined;
  isChatOwner?: boolean | undefined;
  isChatSponsor?: boolean | undefined;
  isChatModerator?: boolean | undefined;
}

export interface LiveChatMessageSnippet {
  type?: LiveChatMessageSnippet_TypeWrapper_Type | undefined;
  liveChatId?: string | undefined;
  authorChannelId?: string | undefined;
  publishedAt?: string | undefined;
  hasDisplayContent?: boolean | undefined;
  displayMessage?: string | undefined;
  textMessageDetails?: LiveChatTextMessageDetails | undefined;
  messageDeletedDetails?: LiveChatMessageDeletedDetails | undefined;
  messageRetractedDetails?: LiveChatMessageRetractedDetails | undefined;
  userBannedDetails?: LiveChatUserBannedMessageDetails | undefined;
  superChatDetails?: LiveChatSuperChatDetails | undefined;
  superStickerDetails?: LiveChatSuperStickerDetails | undefined;
  newSponsorDetails?: LiveChatNewSponsorDetails | undefined;
  memberMilestoneChatDetails?: LiveChatMemberMilestoneChatDetails | undefined;
  membershipGiftingDetails?: LiveChatMembershipGiftingDetails | undefined;
  giftMembershipReceivedDetails?: LiveChatGiftMembershipReceivedDetails | undefined;
  pollDetails?: LiveChatPollDetails | undefined;
}

export interface LiveChatMessageSnippet_TypeWrapper {
}

export enum LiveChatMessageSnippet_TypeWrapper_Type {
  INVALID_TYPE = 0,
  TEXT_MESSAGE_EVENT = 1,
  TOMBSTONE = 2,
  FAN_FUNDING_EVENT = 3,
  CHAT_ENDED_EVENT = 4,
  SPONSOR_ONLY_MODE_STARTED_EVENT = 5,
  SPONSOR_ONLY_MODE_ENDED_EVENT = 6,
  NEW_SPONSOR_EVENT = 7,
  MEMBER_MILESTONE_CHAT_EVENT = 17,
  MEMBERSHIP_GIFTING_EVENT = 18,
  GIFT_MEMBERSHIP_RECEIVED_EVENT = 19,
  MESSAGE_DELETED_EVENT = 8,
  MESSAGE_RETRACTED_EVENT = 9,
  USER_BANNED_EVENT = 10,
  SUPER_CHAT_EVENT = 15,
  SUPER_STICKER_EVENT = 16,
  POLL_EVENT = 20,
  UNRECOGNIZED = -1,
}
// ...
export interface LiveChatTextMessageDetails {
  messageText?: string | undefined;
}

export interface LiveChatMessageDeletedDetails {
  deletedMessageId?: string | undefined;
}

export interface LiveChatMessageRetractedDetails {
  retractedMessageId?: string | undefined;
}

export interface LiveChatUserBannedMessageDetails {
  bannedUserDetails?: ChannelProfileDetails | undefined;
  banType?: LiveChatUserBannedMessageDetails_BanTypeWrapper_BanType | undefined;
  banDurationSeconds?: string | undefined;
}
// ...
export interface LiveChatSuperChatDetails {
  amountMicros?: string | undefined;
  currency?: string | undefined;
  amountDisplayString?: string | undefined;
  userComment?: string | undefined;
  tier?: number | undefined;
}

export interface LiveChatSuperStickerDetails {
  amountMicros?: string | undefined;
  currency?: string | undefined;
  amountDisplayString?: string | undefined;
  tier?: number | undefined;
  superStickerMetadata?: SuperStickerMetadata | undefined;
}

export interface LiveChatFanFundingEventDetails {
  amountMicros?: string | undefined;
  currency?: string | undefined;
  amountDisplayString?: string | undefined;
  userComment?: string | undefined;
}

export interface LiveChatNewSponsorDetails {
  memberLevelName?: string | undefined;
  isUpgrade?: boolean | undefined;
}
// ...
export const LiveChatMessage: MessageFns<LiveChatMessage> = {
  encode(message: LiveChatMessage, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.kind !== undefined && message.kind !== "") {
      writer.uint32(1602).string(message.kind);
    }
    if (message.etag !== undefined && message.etag !== "") {
      writer.uint32(1610).string(message.etag);
    }
    if (message.id !== undefined && message.id !== "") {
      writer.uint32(810).string(message.id);
    }
    if (message.snippet !== undefined) {
      LiveChatMessageSnippet.encode(message.snippet, writer.uint32(18).fork()).join();
    }
    if (message.authorDetails !== undefined) {
      LiveChatMessageAuthorDetails.encode(message.authorDetails, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): LiveChatMessage {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLiveChatMessage();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 200: {
          if (tag !== 1602) {
            break;
          }

          message.kind = reader.string();
          continue;
        }
        case 201: {
          if (tag !== 1610) {
            break;
          }

          message.etag = reader.string();
          continue;
        }
        case 101: {
          if (tag !== 810) {
            break;
          }

          message.id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.snippet = LiveChatMessageSnippet.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.authorDetails = LiveChatMessageAuthorDetails.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): LiveChatMessage {
    return {
      kind: isSet(object.kind) ? globalThis.String(object.kind) : "",
      etag: isSet(object.etag) ? globalThis.String(object.etag) : "",
      id: isSet(object.id) ? globalThis.String(object.id) : "",
      snippet: isSet(object.snippet) ? LiveChatMessageSnippet.fromJSON(object.snippet) : undefined,
      authorDetails: isSet(object.authorDetails)
        ? LiveChatMessageAuthorDetails.fromJSON(object.authorDetails)
        : undefined,
    };
  },

  toJSON(message: LiveChatMessage): unknown {
    const obj: any = {};
    if (message.kind !== undefined && message.kind !== "") {
      obj.kind = message.kind;
    }
    if (message.etag !== undefined && message.etag !== "") {
      obj.etag = message.etag;
    }
    if (message.id !== undefined && message.id !== "") {
      obj.id = message.id;
    }
    if (message.snippet !== undefined) {
      obj.snippet = LiveChatMessageSnippet.toJSON(message.snippet);
    }
    if (message.authorDetails !== undefined) {
      obj.authorDetails = LiveChatMessageAuthorDetails.toJSON(message.authorDetails);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<LiveChatMessage>, I>>(base?: I): LiveChatMessage {
    return LiveChatMessage.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<LiveChatMessage>, I>>(object: I): LiveChatMessage {
    const message = createBaseLiveChatMessage();
    message.kind = object.kind ?? "";
    message.etag = object.etag ?? "";
    message.id = object.id ?? "";
    message.snippet = (object.snippet !== undefined && object.snippet !== null)
      ? LiveChatMessageSnippet.fromPartial(object.snippet)
      : undefined;
    message.authorDetails = (object.authorDetails !== undefined && object.authorDetails !== null)
      ? LiveChatMessageAuthorDetails.fromPartial(object.authorDetails)
      : undefined;
    return message;
  },
};
// ...
export const LiveChatTextMessageDetails: MessageFns<LiveChatTextMessageDetails> = {
  encode(message: LiveChatTextMessageDetails, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.messageText !== undefined && message.messageText !== "") {
      writer.uint32(10).string(message.messageText);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): LiveChatTextMessageDetails {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLiveChatTextMessageDetails();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.messageText = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): LiveChatTextMessageDetails {
    return { messageText: isSet(object.messageText) ? globalThis.String(object.messageText) : "" };
  },

  toJSON(message: LiveChatTextMessageDetails): unknown {
    const obj: any = {};
    if (message.messageText !== undefined && message.messageText !== "") {
      obj.messageText = message.messageText;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<LiveChatTextMessageDetails>, I>>(base?: I): LiveChatTextMessageDetails {
    return LiveChatTextMessageDetails.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<LiveChatTextMessageDetails>, I>>(object: I): LiveChatTextMessageDetails {
    const message = createBaseLiveChatTextMessageDetails();
    message.messageText = object.messageText ?? "";
    return message;
  },
};

function createBaseLiveChatMessageDeletedDetails(): LiveChatMessageDeletedDetails {
  return { deletedMessageId: "" };
}

export const LiveChatMessageDeletedDetails: MessageFns<LiveChatMessageDeletedDetails> = {
  encode(message: LiveChatMessageDeletedDetails, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.deletedMessageId !== undefined && message.deletedMessageId !== "") {
      writer.uint32(810).string(message.deletedMessageId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): LiveChatMessageDeletedDetails {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLiveChatMessageDeletedDetails();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 101: {
          if (tag !== 810) {
            break;
          }

          message.deletedMessageId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): LiveChatMessageDeletedDetails {
    return { deletedMessageId: isSet(object.deletedMessageId) ? globalThis.String(object.deletedMessageId) : "" };
  },

  toJSON(message: LiveChatMessageDeletedDetails): unknown {
    const obj: any = {};
    if (message.deletedMessageId !== undefined && message.deletedMessageId !== "") {
      obj.deletedMessageId = message.deletedMessageId;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<LiveChatMessageDeletedDetails>, I>>(base?: I): LiveChatMessageDeletedDetails {
    return LiveChatMessageDeletedDetails.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<LiveChatMessageDeletedDetails>, I>>(
    object: I,
  ): LiveChatMessageDeletedDetails {
    const message = createBaseLiveChatMessageDeletedDetails();
    message.deletedMessageId = object.deletedMessageId ?? "";
    return message;
  },
};
// ...
export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}